requirements.txt
absl-py==2.2.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
dnspython==2.7.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
joblib==1.4.2
keras==3.9.2
libclang==18.1.1
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==1.0.0
namex==0.0.9
numpy==1.26.4
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
paseto==2.0.0
pendulum==3.1.0
pillow==11.2.1
protobuf==4.25.7
pycparser==2.22
pycryptodomex==3.22.0
Pygments==2.19.1
pymongo==4.12.0
pysodium==0.7.18
python-dateutil==2.9.0.post0
requests==2.32.3
rich==14.0.0
six==1.17.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
termcolor==3.0.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
web3
eth-account
imagehash
